@@ -1,8 +1,8 @@
 package com.baeldung.spring.insertableonly.repositorycheck;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
 
 @Entity
 public class RepositoryCheckBook {
