@@ -1,43 +1,36 @@
 package com.baeldung.keystore;
 
+import org.bouncycastle.jce.provider.BouncyCastleProvider;
+import org.bouncycastle.operator.OperatorCreationException;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import sun.security.x509.AlgorithmId;
-import sun.security.x509.CertificateAlgorithmId;
-import sun.security.x509.CertificateSerialNumber;
-import sun.security.x509.CertificateValidity;
-import sun.security.x509.CertificateVersion;
-import sun.security.x509.CertificateX509Key;
-import sun.security.x509.SubjectAlternativeNameExtension;
-import sun.security.x509.X500Name;
-import sun.security.x509.X509CertImpl;
-import sun.security.x509.X509CertInfo;
-import sun.security.x509.CertificateExtensions;
-import sun.security.x509.GeneralNames;
-import sun.security.x509.GeneralName;
-import sun.security.x509.GeneralNameInterface;
-import sun.security.x509.DNSName;
-import sun.security.x509.IPAddressName;
-import sun.security.util.DerOutputStream;
+import org.bouncycastle.asn1.x500.X500Name;
+import org.bouncycastle.asn1.x509.*;
+import org.bouncycastle.cert.*;
+import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter;
+import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
+import org.bouncycastle.operator.ContentSigner;
+import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 
-import javax.crypto.KeyGenerator;
-import javax.crypto.SecretKey;
-import java.io.IOException;
 import java.math.BigInteger;
-import java.security.InvalidKeyException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.KeyStore;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
-import java.security.SecureRandom;
-import java.security.SignatureException;
+
+import java.security.Security;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.Date;
+
+import javax.crypto.KeyGenerator;
+import javax.crypto.SecretKey;
+
+import java.io.IOException;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.KeyStore;
+import java.security.SecureRandom;
+
 
 /**
  * Created by adi on 4/14/18.
@@ -56,10 +49,14 @@ public class JavaKeyStoreUnitTest {
     private static final String MY_PRIVATE_KEY = "myPrivateKey";
     private static final String MY_CERTIFICATE = "myCertificate";
 
+    private static final String PROVIDER = "BC";
+    private static final int VALIDITY_DAYS = 365 * 50;
+
     @Before
     public void setUp() throws Exception {
         //using java cryptography extension keyStore instead of Keystore.getDefaultType
         keyStore = new JavaKeyStore(KEY_STORE_TYPE, KEYSTORE_PWD, KEYSTORE_NAME);
+        Security.addProvider(new BouncyCastleProvider());
     }
 
     @After
@@ -176,57 +173,25 @@ public class JavaKeyStoreUnitTest {
         Assert.assertTrue(result == null);
     }
 
-    private X509Certificate generateSelfSignedCertificate(KeyPair keyPair) throws CertificateException, IOException, NoSuchProviderException, NoSuchAlgorithmException, InvalidKeyException, SignatureException {
-        X509CertInfo certInfo = new X509CertInfo();
-        // Serial number and version
-        certInfo.set(X509CertInfo.SERIAL_NUMBER, new CertificateSerialNumber(new BigInteger(64, new SecureRandom())));
-        certInfo.set(X509CertInfo.VERSION, new CertificateVersion(CertificateVersion.V3));
+    private X509Certificate generateSelfSignedCertificate(KeyPair keyPair) throws CertificateException, IOException, OperatorCreationException {
 
-        // Subject & Issuer
-        X500Name owner = new X500Name(DN_NAME);
-        certInfo.set(X509CertInfo.SUBJECT, owner);
-        certInfo.set(X509CertInfo.ISSUER, owner);
-
-        // Key and algorithm
-        certInfo.set(X509CertInfo.KEY, new CertificateX509Key(keyPair.getPublic()));
-        AlgorithmId algorithm = new AlgorithmId(AlgorithmId.sha1WithRSAEncryption_oid);
-        certInfo.set(X509CertInfo.ALGORITHM_ID, new CertificateAlgorithmId(algorithm));
-
-        // Validity
+        X500Name issuer = new X500Name(DN_NAME);
+        BigInteger serialNumber = new BigInteger(64, new SecureRandom());
         Date validFrom = new Date();
-        Date validTo = new Date(validFrom.getTime() + 50L * 365L * 24L * 60L * 60L * 1000L); //50 years
-        CertificateValidity validity = new CertificateValidity(validFrom, validTo);
-        certInfo.set(X509CertInfo.VALIDITY, validity);
-        
-        GeneralNameInterface dnsName = new DNSName("baeldung.com");
-        DerOutputStream dnsNameOutputStream = new DerOutputStream();
-        dnsName.encode(dnsNameOutputStream);
-        
-        GeneralNameInterface ipAddress = new IPAddressName("127.0.0.1");
-        DerOutputStream ipAddressOutputStream = new DerOutputStream();
-        ipAddress.encode(ipAddressOutputStream);
-        
-        GeneralNames generalNames = new GeneralNames();
-        generalNames.add(new GeneralName(dnsName));
-        generalNames.add(new GeneralName(ipAddress));
-        
-        CertificateExtensions ext = new CertificateExtensions();
-        ext.set(SubjectAlternativeNameExtension.NAME, new SubjectAlternativeNameExtension(generalNames));
+        Date validTo = new Date(validFrom.getTime() + VALIDITY_DAYS * 24L * 60L * 60L * 1000L);
 
-        certInfo.set(X509CertInfo.EXTENSIONS, ext);        
+        X509v3CertificateBuilder certBuilder = new JcaX509v3CertificateBuilder(issuer, serialNumber, validFrom, validTo, issuer, keyPair.getPublic());
 
-        // Create certificate and sign it
-        X509CertImpl cert = new X509CertImpl(certInfo);
-        cert.sign(keyPair.getPrivate(), SHA1WITHRSA);
+        GeneralNames subjectAltNames = new GeneralNames(
+            new GeneralName[] { new GeneralName(GeneralName.dNSName, "baeldung.com"), new GeneralName(GeneralName.iPAddress, "127.0.0.1") });
 
-        // Since the SHA1withRSA provider may have a different algorithm ID to what we think it should be,
-        // we need to reset the algorithm ID, and resign the certificate
-        AlgorithmId actualAlgorithm = (AlgorithmId) cert.get(X509CertImpl.SIG_ALG);
-        certInfo.set(CertificateAlgorithmId.NAME + "." + CertificateAlgorithmId.ALGORITHM, actualAlgorithm);
-        X509CertImpl newCert = new X509CertImpl(certInfo);
-        newCert.sign(keyPair.getPrivate(), SHA1WITHRSA);
+        certBuilder.addExtension(Extension.subjectAlternativeName, false, subjectAltNames);
 
-        return newCert;
+        ContentSigner signer = new JcaContentSignerBuilder(SHA1WITHRSA).setProvider(PROVIDER)
+            .build(keyPair.getPrivate());
+
+        return new JcaX509CertificateConverter().setProvider(PROVIDER)
+            .getCertificate(certBuilder.build(signer));
     }
 
 }
