@@ -1,18 +1,41 @@
 package com.baeldung.logallrequests;
 
-import java.io.CharArrayWriter;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 
+import jakarta.servlet.ServletOutputStream;
+import jakarta.servlet.WriteListener;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpServletResponseWrapper;
 
 public class ResponseWrapper extends HttpServletResponseWrapper {
 
-    private final CharArrayWriter charArrayWriter = new CharArrayWriter();
-    private final PrintWriter writer = new PrintWriter(charArrayWriter);
+    private final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+    private final PrintWriter writer = new PrintWriter(new OutputStreamWriter(outputStream));
 
     public ResponseWrapper(HttpServletResponse response) {
         super(response);
+    }
+
+    @Override
+    public ServletOutputStream getOutputStream() {
+        return new ServletOutputStream() {
+            @Override
+            public boolean isReady() {
+                return true;
+            }
+
+            @Override
+            public void setWriteListener(WriteListener writeListener) {
+            }
+
+            @Override
+            public void write(int b) {
+                outputStream.write(b);
+            }
+        };
     }
 
     @Override
@@ -20,7 +43,14 @@ public class ResponseWrapper extends HttpServletResponseWrapper {
         return writer;
     }
 
+    @Override
+    public void flushBuffer() throws IOException {
+        super.flushBuffer();
+        writer.flush();
+    }
+
     public String getBodyAsString() {
-        return charArrayWriter.toString();
+        writer.flush();
+        return outputStream.toString();
     }
 }
