@@ -1,11 +1,11 @@
 package com.baeldung.spring.data.persistence.truncate;
 
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import javax.transaction.Transactional;
 
 @Repository
 public interface MyEntityRepository extends CrudRepository<MyEntity, Long> {
