@@ -1,9 +1,27 @@
 package com.baeldung.spring.ai.om;
 
+import org.springframework.ai.autoconfigure.anthropic.AnthropicAutoConfiguration;
+import org.springframework.ai.autoconfigure.bedrock.converse.BedrockConverseProxyChatAutoConfiguration;
+import org.springframework.ai.autoconfigure.ollama.OllamaAutoConfiguration;
+import org.springframework.ai.autoconfigure.vectorstore.chroma.ChromaVectorStoreAutoConfiguration;
+import org.springframework.ai.autoconfigure.vectorstore.pgvector.PgVectorStoreAutoConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.PropertySource;
 
-@SpringBootApplication
+/**
+ * Excluding the below auto-configurations to avoid start up
+ * failure. Their corresponding starters are present on the classpath but are
+ * only needed by other articles in the shared codebase.
+ */
+@SpringBootApplication(exclude = {
+    OllamaAutoConfiguration.class,
+    AnthropicAutoConfiguration.class,
+    PgVectorStoreAutoConfiguration.class,
+    ChromaVectorStoreAutoConfiguration.class,
+    BedrockConverseProxyChatAutoConfiguration.class
+})
+@PropertySource("classpath:application-aiassistant.properties")
 public class OrderManagementApplication {
 
 	public static void main(String[] args) {
