@@ -1,7 +1,10 @@
 package com.baeldung.springai.huggingface;
 
+import org.springframework.ai.autoconfigure.anthropic.AnthropicAutoConfiguration;
+import org.springframework.ai.autoconfigure.bedrock.converse.BedrockConverseProxyChatAutoConfiguration;
 import org.springframework.ai.autoconfigure.openai.OpenAiAutoConfiguration;
 import org.springframework.ai.autoconfigure.vectorstore.chroma.ChromaVectorStoreAutoConfiguration;
+import org.springframework.ai.autoconfigure.vectorstore.pgvector.PgVectorStoreAutoConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.PropertySource;
@@ -13,7 +16,10 @@ import org.springframework.context.annotation.PropertySource;
  */
 @SpringBootApplication(exclude = {
     OpenAiAutoConfiguration.class,
-    ChromaVectorStoreAutoConfiguration.class
+    AnthropicAutoConfiguration.class,
+    PgVectorStoreAutoConfiguration.class,
+    ChromaVectorStoreAutoConfiguration.class,
+    BedrockConverseProxyChatAutoConfiguration.class
 })
 @PropertySource("classpath:application-huggingface.properties")
 public class Application {
