@@ -7,6 +7,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import jakarta.transaction.Transactional;
+
+@Transactional
 @SpringBootTest
 public class RecordsAsJpaIntegrationTest {
     @Autowired
@@ -14,9 +17,11 @@ public class RecordsAsJpaIntegrationTest {
 
     @BeforeEach
     void setUp() {
-        Book book = new Book(1L,"The Lord of the Rings", "J.R.R. Tolkien", "978-0544003415");
-        Book book2 = new Book(2L,"The Hobbit", "J.R.R. Tolkien", "978-0547928227");
-        Book book3 = new Book(3L,"Harry Potter and the Philosopher's Stone", "J.K. Rowling", "978-0747532699");
+
+        Book book = new Book(null,"The Lord of the Rings", "J.R.R. Tolkien", "978-0544003415");
+        Book book2 = new Book(null,"The Hobbit", "J.R.R. Tolkien", "978-0547928227");
+        Book book3 = new Book(null,"Harry Potter and the Philosopher's Stone", "J.K. Rowling", "978-0747532699");
+
 
         bookRepository.save(book);
         bookRepository.save(book2);
