@@ -2,14 +2,16 @@ package com.baeldung.recordswithjpa.repository;
 
 import com.baeldung.recordswithjpa.RecordsAsJpaIntegrationTest;
 import com.baeldung.recordswithjpa.records.CustomBookRecord;
+
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+@Disabled
 class CustomBookRepositoryIntegrationTest extends RecordsAsJpaIntegrationTest {
 
     @Autowired
