@@ -3,17 +3,42 @@ package com.baeldung.config;
 import com.baeldung.contexts.Greeting;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.boot.web.server.WebServerFactoryCustomizer;
+
 import org.springframework.boot.web.servlet.server.ConfigurableServletWebServerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.MediaType;
+
+import org.springframework.web.servlet.View;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.ContentNegotiationConfigurer;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.view.BeanNameViewResolver;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+
 
 /**
  * Web Configuration for the entire app
  */
 @Configuration
 @EnableWebMvc
-public class WebConfig {
+public class WebConfig implements WebMvcConfigurer {
+
+    /**
+     * Spring Boot allows configuring Content Negotiation using properties
+     */
+    @Override
+    public void configureContentNegotiation(final ContentNegotiationConfigurer configurer) {
+        configurer.favorParameter(true)
+            .parameterName("mediaType")
+            .ignoreAcceptHeader(false)
+            .useRegisteredExtensionsOnly(false)
+            .defaultContentType(MediaType.APPLICATION_JSON)
+            .mediaType("xml", MediaType.APPLICATION_XML)
+            .mediaType("json", MediaType.APPLICATION_JSON);
+    }
 
     @Bean
     public WebServerFactoryCustomizer<ConfigurableServletWebServerFactory> enableDefaultServlet() {
@@ -31,4 +56,38 @@ public class WebConfig {
     public ObjectMapper objectMapper() {
         return new ObjectMapper();
     }
+
+    @Bean
+    public ViewResolver viewResolver() {
+        final InternalResourceViewResolver bean = new InternalResourceViewResolver();
+        bean.setViewClass(JstlView.class);
+        bean.setPrefix("/WEB-INF/view/");
+        bean.setSuffix(".jsp");
+        bean.setOrder(2);
+        return bean;
+    }
+
+
+    @Bean
+    public BeanNameViewResolver beanNameViewResolver(){
+        BeanNameViewResolver beanNameViewResolver = new BeanNameViewResolver();
+        beanNameViewResolver.setOrder(1);
+        return beanNameViewResolver;
+    }
+
+    @Bean
+    public View sample() {
+        return new JstlView("/WEB-INF/view/sample.jsp");
+    }
+
+    @Bean
+    public View sample2() {
+        return new JstlView("/WEB-INF/view2/sample2.jsp");
+    }
+
+    @Bean
+    public View sample3(){
+        return new JstlView("/WEB-INF/view3/sample3.jsp");
+    }
+
 }
