@@ -6,7 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = { com.baeldung.web.upload.app.UploadApplication.class, })
+@SpringBootTest(classes = { com.baeldung.compress.SpringCompressRequestApplication.class, })
 public class SpringContextTest {
 
     @Test
