@@ -1,4 +1,4 @@
-package com.baeldung.sharedatasteps;
+package com.baeldung.cucumber.sharedatasteps;
 
 public enum EventStatus {
     PROCESSING, ERROR, COMPLETE
