@@ -1,10 +1,8 @@
-package com.baeldung.sharedatasteps;
+package com.baeldung.cucumber.sharedatasteps;
 
 import org.springframework.boot.test.context.SpringBootTest;
 
-import io.cucumber.spring.CucumberContextConfiguration;
-
-@CucumberContextConfiguration
 @SpringBootTest
 class CucumberIntegrationTest {
+
 }
